pyproject.toml
[project]
name = "papertok-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Manuel Cardenas", email = "manoletre@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "kokoro>=0.9.2",
    "manim>=0.19.0",
    "mistralai>=1.10.1",
    "openai>=2.15.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pip>=25.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "soundfile>=0.12.1",
    "google-genai>=0.2.0",
]

[tool.uv]
package = false
